remove reader.batched and use batch_size as argument for train/test/infer
our reader definition is func that yield a single item,
it should not be a batch item.
reader.batched creates a func that yield a batch items,
which does not belong to decorator and should not be exposed.
